reshuffle code
