+++ cloud/storage/core/libs/grpc/ya.make
@@ -10,6 +10,8 @@ SRCS(
     init.cpp
     keepalive.cpp
     request.cpp
+    periodic_tls_certificate_provider.cpp
+    static_tls_certificate_provider.cpp
     threadpool.cpp
     time_point_specialization.cpp
     tls_certificate_provider.cpp
